refactor(anthropic-direct): extract the round connection phase (4/7)
  loop/round-request.ts     cache breakpoint, params build, connection-phase
                            529/503 retry, watchdog arming, toWireTool
  loop/throttle-signals.ts  racing a parked messages.create against the
                            out-of-band throttle queue

openRound() returns a discriminated OpenRoundResult: `opened` with a live
stream plus both watchdog handles, `retry-ttfb` when the first-byte bound
tripped before any content, or `terminated` when it already yielded the
turn's terminal event.

Watchdog ownership is now explicit and documented: on a successful open the
handles are returned LIVE and disposal transfers to the stream consumer.
Only the failure and TTFB-re-drive paths dispose here. That asymmetry was
implicit in the old inline code and is the reason neither module wraps
disposal in a finally.

throttle-signals.ts is its own file rather than part of round-request
because it is a self-contained concurrency concern with no knowledge of
rounds, retries, or turns — and because folding it in would have pushed
round-request to ~360 lines, over the house ceiling.

loop.ts: 898 -> 649 lines.

Gates: pnpm lint clean; 526 tests pass across 51 files, unchanged.

Refs #360
